w3f_bls::chaum_pedersen_signature

Trait ChaumPedersenSigner

pub trait ChaumPedersenSigner<E: EngineBLS, H: DynDigest + Default + Clone> {
    // Required methods
    fn generate_cp_signature(
        &mut self,
        message: &Message,
    ) -> ChaumPedersenSignature<E>;
    fn generate_witness_scaler(
        &self,
        message_point_as_bytes: &Vec<u8>,
    ) -> <<E as EngineBLS>::PublicKeyGroup as Group>::ScalarField;
    fn generate_dleq_proof(
        &mut self,
        message: &Message,
        bls_signature: E::SignatureGroup,
    ) -> SchnorrProof<E>;
}

ProofOfPossion trait which should be implemented by secret

Required Methods

fn generate_cp_signature( &mut self, message: &Message, ) -> ChaumPedersenSignature<E>

The proof of possession generator is supposed to to produce a schnoor signature of the message using the secret key which it claim to possess.

fn generate_witness_scaler( &self, message_point_as_bytes: &Vec<u8>, ) -> <<E as EngineBLS>::PublicKeyGroup as Group>::ScalarField

fn generate_dleq_proof( &mut self, message: &Message, bls_signature: E::SignatureGroup, ) -> SchnorrProof<E>