w3f_bls

Module double

BLS key pair with public key in both G1 and G2

Unaggreagated BLS signature along side with their DLEQ proof

Implements schemes suggested the paper

The scheme proposes for the public key be represented by doube points, both in G1 and G2 and aggregate keys in G1.

It also proposes that each individual BLS signature accompany a DLEQ proof for faster verification

Structs

DoublePublicKey

BLS Public Key with sub keys in both groups.

DoubleSignature

Detached BLS Signature containing DLEQ

DoubleSignedMessage

Message with attached BLS signature

PublicKeyInSignatureGroup

Wrapper for a point in the signature group which is supposed to the same logarithm as the public key in the public key group

Traits

DoublePublicKeyScheme