w3f_bls::engine

Struct TinyBLS

pub struct TinyBLS<E: Pairing, P: Bls12Config + CurveExtraConfig>(pub E, _)
where
    <P as Bls12Config>::G1Config: WBConfig,
    WBMap<<P as Bls12Config>::G1Config>: MapToCurve<<E as Pairing>::G1>;

Infrequently used BLS variant with tiny 48 byte signatures and 96 byte public keys,

We recommend gainst this variant by default because verifiers always perform O(signers) additions on the PublicKeyGroup, or worse 128 bit scalar multiplications with delinearization. Yet, there are specific use cases where this variant performs better. We swapy two group roles relative to zcash here.

Tuple Fields
0: E