w3f_bls

Trait ProofOfPossessionGenerator

pub trait ProofOfPossessionGenerator<E: EngineBLS, H: DynDigest + Default + Clone, PV, P: ProofOfPossession<E, H, PV>> {
    // Required method
    fn generate_pok(&mut self) -> P;
}

ProofOfPossion trait which should be implemented by secret

Required Methods

fn generate_pok(&mut self) -> P

The proof of possession generator is supposed to to produce a schnorr signature or a bls signature using the secret key which it claims to possess. This proves that that the secret key is known.