w3f_bls::verifiers

Function verify_simple

pub fn verify_simple<S: Signed>(s: S) -> bool

Simple universal BLS signature verification

We support an unstable Signed::messages_and_publickeys() securely by calling it only once and batch normalizing all points, as do most other verification routines here. We do no optimizations that reduce the number of pairings by combining repeated messages or signers.